+++ src/Adapter/OpenIdConnectRestApiProxy/OPEN_ID_CONNECT_REST_API_PROXY.mjs
@@ -1,0 +1,3 @@
+import { PROTOCOL_HTTP } from "../../../../flux-http-api/src/Adapter/Protocol/PROTOCOL.mjs";
+
+export const OPEN_ID_CONNECT_REST_API_PROXY_DEFAULT_OPEN_ID_CONNECT_REST_API_URL = `${PROTOCOL_HTTP}://open-id-connect-rest-api:9501`;